Runtime deviations checking (#30)
* Runtime deviations checking

+++ perf/scripts/gen_website.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+if [ $# -ne 1 ]; then
+   echo "Usage: $0 <datafolder>. Exiting."
+   exit 1
+fi
 
 rm -rf tmp && mkdir tmp
 cd tmp && git clone https://github.com/open-quantum-safe/profiling.git && cd ..
@@ -10,5 +14,20 @@ cp tmp/profiling/visualization/*.js out
 cp tmp/profiling/visualization/*.css out
 cd out && tar czvf ../site.tgz * && cd ..
 cp site.tgz ${1}/site
-rm -rf tmp out
+
+# perform deviation test and reporting
+
+# obtain the last deviation report
+cp $1/devs.txt .
+# create new deviations report
+python3 devcheck.py out devs.txt > newdevs.txt
+# discard cycle counting deviations assuming they're caused by swapping
+egrep -v "insts|cycles" newdevs.txt > report.txt
+# send email
+python3 notify.py report.txt
+# store todays deviation for comparison tomorrow
+cp newdevs.txt $1/devs.txt
+
+# cleanup
+rm -rf tmp out report.txt *devs.txt
